added all intrinsics
On the wiki table, the teleport and polymorph properties are listed as "teleportation/teleportisis"/"polymorph/polymorphitis" so I added them as two different properties but with identical effects/sources - don't know if this was the right thing to do? Also, two telepathy and speed entries: (intrinsic) and (extrinsic). Also another thing you should ask me about and I'll explain